backend/src/requirements.txt
django==4.2.*
djangorestframework==3.15.*
djangorestframework-simplejwt==5.3.*
drf_spectacular~=0.27.2
django-cors-headers~=4.3.1
# django-guardian~=2.4.0
django-oauth-toolkit~=2.3.0
django-debug-toolbar~=4.3.0
whitenoise[brotli]==6.6.0
drf-yasg~=1.21.7
gunicorn==22.0.0
psycopg2_binary==2.9.9
markdown~=3.5.1
Pygments~=2.16.1
PyYAML~=6.0.1
pillow~=10.1.0
openai==1.76.0
langchain~=0.1.19
langchain-community~=0.0.38
ruff~=0.4.4
pre-commit==3.5.0
pypdf~=4.2.0
chromadb~=0.5.0
tiktoken~=0.6.0
celery[redis]~=5.4.0
flower~=2.0.1
unstructured~=0.13.7
termcolor~=2.4.0
channels[daphne]==4.1.0
channels-redis==4.2.0
